agregado subcategorias a tabla categories

+++ app/Http/Controllers/User/UserController.php
@@ -140,7 +140,7 @@ class UserController extends ApiController
             $user->avatar = $request->avatar;
         }
 
-        /*if ($request->has("photo")) {
+        if ($request->has("photo")) {
             if ($user->photo)
                 Storage::delete($user->photo);
             $image = $request->photo;
@@ -149,7 +149,7 @@ class UserController extends ApiController
             $imageName = Str::random(10) . '.jpeg';
             Storage::disk('photos')->put($imageName, base64_decode($image));
             $user->photo =  $imageName;
-        }*/
+        }
 
         if ($request->has('photo')) {
             $user->photo = $request->photo;
@@ -162,6 +162,7 @@ class UserController extends ApiController
         if ($request->has("surveyed")) {
             $user->surveyed = $request->surveyed;
         }
+
         if ($request->has("premium")) {
             $user->premium = $request->premium;
         }
@@ -181,7 +182,6 @@ class UserController extends ApiController
         if ($request->has("fortaleza_mental")) {
             $user->fortaleza_mental = $request->fortaleza_mental;
         }
-
 
         if (!$user->isDirty()) {
             return $this->errorResponse(
